requirements.txt
bonfires>=0.4.0
httpx>=0.27
python-dotenv>=1.0
pytest>=8.0
